+++ NEWS.md
@@ -1,3 +1,4 @@
+- added cc_and() and cc_or() as useful wrappers
 - set domain = NA for messages, warnings and errors
 
 # err 0.0.1